+++ lib/brando/query/mutations.ex
@@ -64,6 +64,42 @@ defmodule Brando.Query.Mutations do
     end
   end
 
+  def duplicate(context, _module, name, id, opts, user) do
+    {:ok, entry} = apply(context, :"get_#{name}", [id])
+    opts = Enum.into(opts, %{})
+
+    params =
+      entry
+      |> maybe_change_fields(opts)
+      |> maybe_delete_fields(opts)
+      |> maybe_set_status()
+      |> Map.from_struct()
+      |> drop_id()
+
+    apply(context, :"create_#{name}", [params, user])
+  end
+
+  defp drop_id(%{id: _} = entry), do: Map.drop(entry, [:id])
+  defp drop_id(entry), do: entry
+
+  defp maybe_set_status(%{status: _} = entry), do: Map.put(entry, :status, :draft)
+  defp maybe_set_status(entry), do: entry
+
+  defp maybe_change_fields(entry, %{change_fields: change_fields}) do
+    Enum.reduce(change_fields, entry, fn f, updated_entry ->
+      default_value = Map.get(updated_entry, f)
+      Map.update(updated_entry, f, default_value, fn v -> "#{v}_dupl" end)
+    end)
+  end
+
+  defp maybe_change_fields(entry, _), do: entry
+
+  defp maybe_delete_fields(entry, %{delete_fields: delete_fields}) do
+    Map.delete(entry, delete_fields)
+  end
+
+  defp maybe_delete_fields(entry, _), do: entry
+
   def delete(context, module, name, id, user, callback_block) do
     {:ok, entry} = apply(context, :"get_#{name}", [id])